# Implement better booster logging