--- a/booster/library/Booster/CLOptions.hs
+++ b/booster/library/Booster/CLOptions.hs
@@ -30,6 +30,8 @@ data CLOptions = CLOptions
     , llvmLibraryFile :: Maybe FilePath
     , port :: Int
     , logLevels :: [LogLevel]
+    , logContexts :: [String]
+    , notLogContexts :: [String]
     , simplificationLogFile :: Maybe FilePath
     , smtOptions :: Maybe SMTOptions
     , equationOptions :: EquationOptions
@@ -76,6 +78,25 @@ clOptionsParser =
                           \or a custom level: "
                             <> intercalate ", " (map fst allowedLogLevels)
                         )
+                )
+            )
+        <*> many
+            ( option
+                str
+                ( metavar "CONTEXT"
+                    <> long "log-context"
+                    <> short 'c'
+                    <> help
+                        "Log context"
+                )
+            )
+        <*> many
+            ( option
+                str
+                ( metavar "CONTEXT"
+                    <> long "not-log-context"
+                    <> help
+                        "Not in log context"
                 )
             )
         <*> optional