--- a/booster/library/Booster/Definition/Base.hs
+++ b/booster/library/Booster/Definition/Base.hs
@@ -32,6 +32,7 @@ import Booster.Log (ToLogFormat (..))
 import GHC.TypeLits (KnownSymbol, symbolVal)
 import Data.Data (Proxy(..))
 import GHC.Base (coerce)
+import Data.Text.Lazy qualified as LazyText
 
 {- | A Kore definition is constructed from a main module with its
    transitive imports.
@@ -111,7 +112,7 @@ instance Pretty SourceRef where
         UNKNOWN -> "UNKNOWN"
 
 instance KnownSymbol k => ToLogFormat (RewriteRule k) where
-    toTextualLog RewriteRule{attributes} = pack (symbolVal (Proxy :: Proxy k)) <> " " <> maybe "UNKNWON" coerce attributes.uniqueId
+    toTextualLog RewriteRule{attributes} = LazyText.toStrict $ (LazyText.toLower $ LazyText.pack $ symbolVal (Proxy :: Proxy k)) <> " " <> maybe "UNKNOWN" (LazyText.take 7 . LazyText.fromStrict . coerce) attributes.uniqueId
 
 -- | class of entities that have a location or ID to present to users
 class HasSourceRef a where