--- a/booster/library/Booster/Definition/Base.hs
+++ b/booster/library/Booster/Definition/Base.hs
@@ -25,13 +25,13 @@ import Booster.Definition.Attributes.Base
 import Booster.Pattern.Base
 import Booster.Pattern.Index (TermIndex)
 import Control.Applicative ((<|>))
+import Data.Data (Proxy (..))
 import Data.Maybe (fromMaybe)
 import Data.Text (Text, pack)
-import Prettyprinter (Pretty (..))
-import GHC.TypeLits (KnownSymbol, symbolVal)
-import Data.Data (Proxy(..))
-import GHC.Base (coerce)
 import Data.Text.Lazy qualified as LazyText
+import GHC.Base (coerce)
+import GHC.TypeLits (KnownSymbol, symbolVal)
+import Prettyprinter (Pretty (..))
 
 {- | A Kore definition is constructed from a main module with its
    transitive imports.