--- a/booster/library/Booster/Log.hs
+++ b/booster/library/Booster/Log.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 
 module Booster.Log (module Booster.Log) where