… booster/library/Booster/Pattern/Base.hs
… booster/library/Booster/Pattern/Match.hs
… booster/library/Booster/SMT/Interface.hs
… booster/library/Booster/SMT/Runner.hs
… booster/library/Booster/Syntax/Json/Internalise.hs
… booster/package.yaml
… booster/test/rpc-integration/runDirectoryTest.sh
… booster/test/rpc-integration/test-log-simplify-json/test.sh
… booster/tools/booster/Server.hs
… dev-tools/booster-dev/Server.hs
… dev-tools/package.yaml
… docs/logging.md