This PR structured logs which allow us to more easily trace rewriting and simplifications:

```
[Info#proxy] Starting execute request
[booster][execute][smt] Starting new SMT solver
[booster][execute][smt] Solver ready to use
[booster][execute][smt] Checking definition prelude
[booster][execute][term 58b79a1][detail] <generatedTop>(<foundry>(<kevm>(<k>(kseq(#execute_EVM_KItem(), VarCONTINUATION:SortK{})), <exit-code>(VarEXITCODE_CELL:SortInt{}), <mode>(NORMAL()), <schedule>(SHANGHAI_EVM()), <useGas>("false"), <ethereum>(<evm>(<output>(VarOUTPUT_CELL:SortBytes{}), <statusCode>(VarSTATUSCODE:SortStatusCode{}), <callStack>(VarCALLSTACK_CELL:SortList{}), <interimStates>(VarINTERIMSTATES_CELL:SortList{}), <touchedAccounts>(VarTOUCHEDACCOUNTS_CELL:SortSet{}), <callState>(<program>("`\128`@R4\128\NAKa\NUL\DLEW`\NUL\128\253...truncated"), <jumpDests>({"16", "87", "92", "353", "344", "365", "332", "397", "381", "296", "217", "224", "200", "205", "111", "106", "167", "162", "129", "181", "186", "143", "148", "745", "720", "618", "623", "600", "653", "686", "680", "593", "570", "575", "518", "551", "529", "452", "475", "494", "416", "431"} ), <id>(VarCONTRACT_ID:SortInt{}), <caller>(VarCALLER_ID:SortInt{}), <callData>(_+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes("\156&\224\&7", _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV0_x_114b9705:SortInt{}), _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV1_y_114b9705:SortInt{}), #buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV2_z_114b9705:SortInt{}))))), <callValue>("0"), <wordStack>(.WordStack_EVM-TYPES_WordStack()), <localMem>(""), <pc>("0"), <gas>("0"), <memoryUsed>("0"), <callGas>("0"), <static>(VarSTATIC_CELL:SortBool{}), <callDepth>(VarCALLDEPTH_CELL:SortInt{})), <substate>(<selfDestruct>(VarSELFDESTRUCT_CELL:SortSet{}), <log>(VarLOG_CELL:SortList{}), <refund>("0"), <accessedAccounts>(VarACCESSEDACCOUNTS_CELL:SortSet{}), <accessedStorage>(VarACCESSEDSTORAGE_CELL:SortMap{})), <gasPrice>(VarGASPRICE_CELL:SortInt{}), <origin>(VarORIGIN_ID:SortInt{}), <blockhashes>(VarBLOCKHASHES_CELL:SortList{}), <block>(<previousHash>(VarPREVIOUSHASH_CELL:SortInt{}), <ommersHash>(VarOMMERSHASH_CELL:SortInt{}), <coinbase>(VarCOINBASE_CELL:SortInt{}), <stateRoot>(VarSTATEROOT_CELL:SortInt{}), <transactionsRoot>(VarTRANSACTIONSROOT_CELL:SortInt{}), <receiptsRoot>(VarRECEIPTSROOT_CELL:SortInt{}), <logsBloom>(VarLOGSBLOOM_CELL:SortBytes{}), <difficulty>(VarDIFFICULTY_CELL:SortInt{}), <number>(VarNUMBER_CELL:SortInt{}), <gasLimit>(VarGASLIMIT_CELL:SortInt{}), <gasUsed>(VarGASUSED_CELL:SortGas{}), <timestamp>(VarTIMESTAMP_CELL:SortInt{}), <extraData>(VarEXTRADATA_CELL:SortBytes{}), <mixHash>(VarMIXHASH_CELL:SortInt{}), <blockNonce>(VarBLOCKNONCE_CELL:SortInt{}), <baseFee>(VarBASEFEE_CELL:SortInt{}), <withdrawalsRoot>(VarWITHDRAWALSROOT_CELL:SortInt{}), <ommerBlockHeaders>(VarOMMERBLOCKHEADERS_CELL:SortJSON{}))), <network>(<chainID>(VarCHAINID_CELL:SortInt{}), <accounts>({<acctID>(VarCONTRACT_ID:SortInt{}) -> <account>(<acctID>(VarCONTRACT_ID:SortInt{}), <balance>(VarCONTRACT_BAL:SortInt{}), <code>("`\128`@R4\128\NAKa\NUL\DLEW`\NUL\128\253...truncated"), <storage>(VarCONTRACT_STORAGE:SortMap{}), <origStorage>(VarCONTRACT_ORIGSTORAGE:SortMap{}), <nonce>(VarCONTRACT_NONCE:SortInt{})), VarACCOUNTS_REST:SortAccountCellMap{}}), <txOrder>(VarTXORDER_CELL:SortList{}), <txPending>(VarTXPENDING_CELL:SortList{}), <messages>(VarMESSAGES_CELL:SortMessageCellMap{})))), <cheatcodes>(<prank>(<prevCaller>(VarPREVCALLER_CELL:SortAccount{}), <prevOrigin>(VarPREVORIGIN_CELL:SortAccount{}), <newCaller>(VarNEWCALLER_CELL:SortAccount{}), <newOrigin>(VarNEWORIGIN_CELL:SortAccount{}), <active>("false"), <depth>(VarDEPTH_CELL:SortInt{}), <singleCall>("false")), <expectedRevert>(<isRevertExpected>("false"), <expectedReason>(VarEXPECTEDREASON_CELL:SortBytes{}), <expectedDepth>(VarEXPECTEDDEPTH_CELL:SortInt{})), <expectedOpcode>(<isOpcodeExpected>("false"), <expectedAddress>(VarEXPECTEDADDRESS_CELL:SortAccount{}), <expectedValue>(VarEXPECTEDVALUE_CELL:SortInt{}), <expectedData>(VarEXPECTEDDATA_CELL:SortBytes{}), <opcodeType>(VarOPCODETYPE_CELL:SortOpcodeType{})), <expectEmit>(<recordEvent>("false"), <isEventExpected>("false"), <checkedTopics>(VarCHECKEDTOPICS_CELL:SortList{}), <checkedData>(VarCHECKEDDATA_CELL:SortBool{}), <expectedEventAddress>(VarEXPECTEDEVENTADDRESS_CELL:SortAccount{})), <whitelist>(<isCallWhitelistActive>("false"), <isStorageWhitelistActive>("false"), <addressSet>({}), <storageSlotSet>({})), <mockCalls>({})), <KEVMTracing>(<activeTracing>("false"), <traceStorage>("false"), <traceWordStack>("false"), <traceMemory>("false"), <recordedTrace>("false"), <traceData>([]))), <generatedCounter>(VarGENERATEDCOUNTER_CELL:SortInt{})) /\ notBool_(AccountCellMap:in_keys(<acctID>(VarCONTRACT_ID:SortInt{}), VarACCOUNTS_REST:SortAccountCellMap{})) /\ _<Int_(VarCALLER_ID:SortInt{}, "1461501637330902...truncated") /\ _<=Int_("0", VarCALLER_ID:SortInt{}) /\ notBool_(_andBool_(_<Int_("0", VarCALLER_ID:SortInt{}), _<=Int_(VarCALLER_ID:SortInt{}, "9"))) /\ _=/=Int_(VarCALLER_ID:SortInt{}, "6453264744265472...truncated") /\ _<=Int_("0", VarCONTRACT_BAL:SortInt{}) /\ _<Int_(VarCONTRACT_BAL:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarCONTRACT_ID:SortInt{}) /\ _<Int_(VarCONTRACT_ID:SortInt{}, "1461501637330902...truncated") /\ _=/=Int_(VarCONTRACT_ID:SortInt{}, "6453264744265472...truncated") /\ notBool_(_andBool_(_<Int_("0", VarCONTRACT_ID:SortInt{}), _<=Int_(VarCONTRACT_ID:SortInt{}, "9"))) /\ _<Int_(VarCONTRACT_NONCE:SortInt{}, "1844674407370955...truncated") /\ _<=Int_("0", VarCONTRACT_NONCE:SortInt{}) /\ _<=Int_(VarNUMBER_CELL:SortInt{}, "5789604461865809...truncated") /\ _<=Int_("0", VarNUMBER_CELL:SortInt{}) /\ _<=Int_("0", VarORIGIN_ID:SortInt{}) /\ _=/=Int_(VarORIGIN_ID:SortInt{}, "6453264744265472...truncated") /\ _<Int_(VarORIGIN_ID:SortInt{}, "1461501637330902...truncated") /\ notBool_(_andBool_(_<Int_("0", VarORIGIN_ID:SortInt{}), _<=Int_(VarORIGIN_ID:SortInt{}, "9"))) /\ _<Int_(VarTIMESTAMP_CELL:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarTIMESTAMP_CELL:SortInt{}) /\ _<Int_(VarVV0_x_114b9705:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarVV0_x_114b9705:SortInt{}) /\ _<Int_(VarVV1_y_114b9705:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarVV1_y_114b9705:SortInt{}) /\ _<Int_(VarVV2_z_114b9705:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarVV2_z_114b9705:SortInt{})
[booster][execute][term 58b79a1][rewrite UNKNOWN][detail] BASIC-BLOCK-8-TO-6
[booster][execute][term 58b79a1][rewrite UNKNOWN][match][abort] Uncertain about match with rule. Remainder: [(_+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes("w\SYN\STX\247", _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", Rule#VarVV0_x_114b9705:SortInt{}), #buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", Rule#VarVV1_y_114b9705:SortInt{}))), _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes("\156&\224\&7", _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV0_x_114b9705:SortInt{}), _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV1_y_114b9705:SortInt{}), #buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV2_z_114b9705:SortInt{}))))), ({"16", "87", "92", "353", "344", "365", "332", "397", "381", "296", "217", "224", "200", "205", "111", "106", "167", "162", "129", "181", "186", "143", "148", "745", "720", "618", "623", "600", "653", "686", "680", "593", "570", "575", "518", "551", "529", "452", "475", "494", "416", "431"} , {"16", "87", "92", "353", "344", "365", "332", "397", "381", "296", "217", "224", "200", "205", "111", "106", "167", "162", "129", "181", "186", "143", "148", "745", "720", "618", "623", "600", "653", "686", "680", "593", "570", "575", "518", "551", "529", "452", "475", "494", "416", "431"} ), ({}, {}), ({}, {})]
[booster][execute][term 58b79a1][detail] <generatedTop>(<foundry>(<kevm>(<k>(kseq(#execute_EVM_KItem(), VarCONTINUATION:SortK{})), <exit-code>(VarEXITCODE_CELL:SortInt{}), <mode>(NORMAL()), <schedule>(SHANGHAI_EVM()), <useGas>("false"), <ethereum>(<evm>(<output>(VarOUTPUT_CELL:SortBytes{}), <statusCode>(VarSTATUSCODE:SortStatusCode{}), <callStack>(VarCALLSTACK_CELL:SortList{}), <interimStates>(VarINTERIMSTATES_CELL:SortList{}), <touchedAccounts>(VarTOUCHEDACCOUNTS_CELL:SortSet{}), <callState>(<program>("`\128`@R4\128\NAKa\NUL\DLEW`\NUL\128\253...truncated"), <jumpDests>({"16", "87", "92", "353", "344", "365", "332", "397", "381", "296", "217", "224", "200", "205", "111", "106", "167", "162", "129", "181", "186", "143", "148", "745", "720", "618", "623", "600", "653", "686", "680", "593", "570", "575", "518", "551", "529", "452", "475", "494", "416", "431"} ), <id>(VarCONTRACT_ID:SortInt{}), <caller>(VarCALLER_ID:SortInt{}), <callData>(_+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes("\156&\224\&7", _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV0_x_114b9705:SortInt{}), _+Bytes__BYTES-HOOKED_Bytes_Bytes_Bytes(#buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV1_y_114b9705:SortInt{}), #buf(_,_)_BUF-SYNTAX_Bytes_Int_Int("32", VarVV2_z_114b9705:SortInt{}))))), <callValue>("0"), <wordStack>(.WordStack_EVM-TYPES_WordStack()), <localMem>(""), <pc>("0"), <gas>("0"), <memoryUsed>("0"), <callGas>("0"), <static>(VarSTATIC_CELL:SortBool{}), <callDepth>(VarCALLDEPTH_CELL:SortInt{})), <substate>(<selfDestruct>(VarSELFDESTRUCT_CELL:SortSet{}), <log>(VarLOG_CELL:SortList{}), <refund>("0"), <accessedAccounts>(VarACCESSEDACCOUNTS_CELL:SortSet{}), <accessedStorage>(VarACCESSEDSTORAGE_CELL:SortMap{})), <gasPrice>(VarGASPRICE_CELL:SortInt{}), <origin>(VarORIGIN_ID:SortInt{}), <blockhashes>(VarBLOCKHASHES_CELL:SortList{}), <block>(<previousHash>(VarPREVIOUSHASH_CELL:SortInt{}), <ommersHash>(VarOMMERSHASH_CELL:SortInt{}), <coinbase>(VarCOINBASE_CELL:SortInt{}), <stateRoot>(VarSTATEROOT_CELL:SortInt{}), <transactionsRoot>(VarTRANSACTIONSROOT_CELL:SortInt{}), <receiptsRoot>(VarRECEIPTSROOT_CELL:SortInt{}), <logsBloom>(VarLOGSBLOOM_CELL:SortBytes{}), <difficulty>(VarDIFFICULTY_CELL:SortInt{}), <number>(VarNUMBER_CELL:SortInt{}), <gasLimit>(VarGASLIMIT_CELL:SortInt{}), <gasUsed>(VarGASUSED_CELL:SortGas{}), <timestamp>(VarTIMESTAMP_CELL:SortInt{}), <extraData>(VarEXTRADATA_CELL:SortBytes{}), <mixHash>(VarMIXHASH_CELL:SortInt{}), <blockNonce>(VarBLOCKNONCE_CELL:SortInt{}), <baseFee>(VarBASEFEE_CELL:SortInt{}), <withdrawalsRoot>(VarWITHDRAWALSROOT_CELL:SortInt{}), <ommerBlockHeaders>(VarOMMERBLOCKHEADERS_CELL:SortJSON{}))), <network>(<chainID>(VarCHAINID_CELL:SortInt{}), <accounts>({<acctID>(VarCONTRACT_ID:SortInt{}) -> <account>(<acctID>(VarCONTRACT_ID:SortInt{}), <balance>(VarCONTRACT_BAL:SortInt{}), <code>("`\128`@R4\128\NAKa\NUL\DLEW`\NUL\128\253...truncated"), <storage>(VarCONTRACT_STORAGE:SortMap{}), <origStorage>(VarCONTRACT_ORIGSTORAGE:SortMap{}), <nonce>(VarCONTRACT_NONCE:SortInt{})), VarACCOUNTS_REST:SortAccountCellMap{}}), <txOrder>(VarTXORDER_CELL:SortList{}), <txPending>(VarTXPENDING_CELL:SortList{}), <messages>(VarMESSAGES_CELL:SortMessageCellMap{})))), <cheatcodes>(<prank>(<prevCaller>(VarPREVCALLER_CELL:SortAccount{}), <prevOrigin>(VarPREVORIGIN_CELL:SortAccount{}), <newCaller>(VarNEWCALLER_CELL:SortAccount{}), <newOrigin>(VarNEWORIGIN_CELL:SortAccount{}), <active>("false"), <depth>(VarDEPTH_CELL:SortInt{}), <singleCall>("false")), <expectedRevert>(<isRevertExpected>("false"), <expectedReason>(VarEXPECTEDREASON_CELL:SortBytes{}), <expectedDepth>(VarEXPECTEDDEPTH_CELL:SortInt{})), <expectedOpcode>(<isOpcodeExpected>("false"), <expectedAddress>(VarEXPECTEDADDRESS_CELL:SortAccount{}), <expectedValue>(VarEXPECTEDVALUE_CELL:SortInt{}), <expectedData>(VarEXPECTEDDATA_CELL:SortBytes{}), <opcodeType>(VarOPCODETYPE_CELL:SortOpcodeType{})), <expectEmit>(<recordEvent>("false"), <isEventExpected>("false"), <checkedTopics>(VarCHECKEDTOPICS_CELL:SortList{}), <checkedData>(VarCHECKEDDATA_CELL:SortBool{}), <expectedEventAddress>(VarEXPECTEDEVENTADDRESS_CELL:SortAccount{})), <whitelist>(<isCallWhitelistActive>("false"), <isStorageWhitelistActive>("false"), <addressSet>({}), <storageSlotSet>({})), <mockCalls>({})), <KEVMTracing>(<activeTracing>("false"), <traceStorage>("false"), <traceWordStack>("false"), <traceMemory>("false"), <recordedTrace>("false"), <traceData>([]))), <generatedCounter>(VarGENERATEDCOUNTER_CELL:SortInt{})) /\ notBool_(AccountCellMap:in_keys(<acctID>(VarCONTRACT_ID:SortInt{}), VarACCOUNTS_REST:SortAccountCellMap{})) /\ _<Int_(VarCALLER_ID:SortInt{}, "1461501637330902...truncated") /\ _<=Int_("0", VarCALLER_ID:SortInt{}) /\ notBool_(_andBool_(_<Int_("0", VarCALLER_ID:SortInt{}), _<=Int_(VarCALLER_ID:SortInt{}, "9"))) /\ _=/=Int_(VarCALLER_ID:SortInt{}, "6453264744265472...truncated") /\ _<=Int_("0", VarCONTRACT_BAL:SortInt{}) /\ _<Int_(VarCONTRACT_BAL:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarCONTRACT_ID:SortInt{}) /\ _<Int_(VarCONTRACT_ID:SortInt{}, "1461501637330902...truncated") /\ _=/=Int_(VarCONTRACT_ID:SortInt{}, "6453264744265472...truncated") /\ notBool_(_andBool_(_<Int_("0", VarCONTRACT_ID:SortInt{}), _<=Int_(VarCONTRACT_ID:SortInt{}, "9"))) /\ _<Int_(VarCONTRACT_NONCE:SortInt{}, "1844674407370955...truncated") /\ _<=Int_("0", VarCONTRACT_NONCE:SortInt{}) /\ _<=Int_(VarNUMBER_CELL:SortInt{}, "5789604461865809...truncated") /\ _<=Int_("0", VarNUMBER_CELL:SortInt{}) /\ _<=Int_("0", VarORIGIN_ID:SortInt{}) /\ _=/=Int_(VarORIGIN_ID:SortInt{}, "6453264744265472...truncated") /\ _<Int_(VarORIGIN_ID:SortInt{}, "1461501637330902...truncated") /\ notBool_(_andBool_(_<Int_("0", VarORIGIN_ID:SortInt{}), _<=Int_(VarORIGIN_ID:SortInt{}, "9"))) /\ _<Int_(VarTIMESTAMP_CELL:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarTIMESTAMP_CELL:SortInt{}) /\ _<Int_(VarVV0_x_114b9705:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarVV0_x_114b9705:SortInt{}) /\ _<Int_(VarVV1_y_114b9705:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarVV1_y_114b9705:SortInt{}) /\ _<Int_(VarVV2_z_114b9705:SortInt{}, "1157920892373161...truncated") /\ _<=Int_("0", VarVV2_z_114b9705:SortInt{})
…