@goodlyrottenapple 2024-04-24
looks like there might be a bit of a slowdown in kevm:

| Test                                                                                      | sam-incremental-logging-update-experiment time | master-089deaf61 time | (sam-incremental-logging-update-experiment/master-089deaf61) time
|-------------------------------------------------------------------------------------------|------------------------------------------------|-----------------------|-------------------------------------------------------------------
| kontrol/test-storetest-teststoreload-0-spec.k                                             | 49.95                                          | 57.63                 | 0.8667360749609578
| kontrol/test-storetest-testaccesses-0-spec.k                                              | 49.94                                          | 56.19                 | 0.8887702438156255
| kontrol/test-expectreverttest-testfail_expectrevert_failandsuccess-0-spec.k               | 58.03                                          | 64.87                 | 0.8945583474641591
| kontrol/test-arithmetictest-test_wdiv_rounding-uint256-uint256-0-spec.k                   | 61.41                                          | 63.95                 | 0.9602814698983579
| bihu/forwardToHotWallet-success-2-spec.k                                                  | 52.48                                          | 50.69                 | 1.0353126849477214
| benchmarks/staticloop00-a0lt10-spec.k                                                     | 48.01                                          | 46.36                 | 1.0355910267471957
| erc20/ds/transferFrom-failure-2-b-spec.k                                                  | 90.92                                          | 87.68                 | 1.0369525547445255
| benchmarks/ecrecover00-siginvalid-spec.k                                                  | 60.09                                          | 57.81                 | 1.0394395433316035
| erc20/hkg/transfer-success-1-spec.k                                                       | 50.93                                          | 48.89                 | 1.041726324401718
| mcd/cat-file-addr-pass-rough-spec.k                                                       | 57.0                                           | 54.6                  | 1.043956043956044
| erc20/ds/transfer-failure-1-b-spec.k                                                      | 91.62                                          | 87.72                 | 1.0444596443228455
| mcd/flipper-bids-pass-rough-spec.k                                                        | 79.3                                           | 75.55                 | 1.0496360026472535
| benchmarks/ecrecover00-sigvalid-spec.k                                                    | 62.05                                          | 58.98                 | 1.052051542895897
| benchmarks/encode-keccak00-spec.k                                                         | 50.29                                          | 47.79                 | 1.0523121992048545
| erc20/ds/transferFrom-failure-1-b-spec.k                                                  | 141.88                                         | 134.67                | 1.0535382787554763
| erc20/hkg/transferFrom-success-1-spec.k                                                   | 70.02                                          | 66.36                 | 1.0551537070524413
| erc20/ds/transferFrom-failure-1-a-spec.k                                                  | 87.26                                          | 82.65                 | 1.0557773744706593
| benchmarks/structarg01-spec.k                                                             | 55.63                                          | 52.44                 | 1.0608314263920673
| erc20/hkg/transferFrom-success-2-spec.k                                                   | 66.61                                          | 62.77                 | 1.0611757208857735
| bihu/forwardToHotWallet-success-1-spec.k                                                  | 60.53                                          | 57.02                 | 1.0615573482988425
| erc20/ds/transfer-success-1-spec.k                                                        | 125.34                                         | 117.93                | 1.0628338845077587
| erc20/ds/approve-success-spec.k                                                           | 63.38                                          | 59.57                 | 1.0639583683061944
| mcd/vat-slip-pass-rough-spec.k                                                            | 107.46                                         | 100.75                | 1.0666004962779156
| erc20/ds/transfer-success-2-spec.k                                                        | 74.25                                          | 68.44                 | 1.0848918760958504
| erc20/ds/transferFrom-success-1-spec.k                                                    | 97.4                                           | 88.76                 | 1.0973411446597567
| erc20/ds/transferFrom-success-2-spec.k                                                    | 85.86                                          | 78.24                 | 1.09739263803681
| kontrol/test-arithmetictest-test_wmul_weakly_increasing_positive-uint256-uint256-0-spec.k | 80.18                                          | 72.41                 | 1.107305620770612
| kontrol/test-expectreverttest-testfail_expectrevert_bytes4-0-spec.k                       | 66.81                                          | 57.57                 | 1.160500260552371
| kontrol/test-allowchangestest-testallow-0-spec.k                                          | 58.76                                          | 50.31                 | 1.1679586563307494
| TOTAL                                                                                     | 2103.39                                        | 2008.6000000000001    | 1.0471920740814495