@JuanCoRo 2024-04-25
It would be great for Kontrol if you could also include the timestamps of each log entry!
This could be toggled if the verbosity is too high or if it causes further slowdown. Having timestamps on this would help us debug https://github.com/runtimeverification/kontrol/issues/526, for instance.

Also, from what I see, can some `[detail]` logs print the entire configuration? If this is the case, can something go wrong when the configuration is enormous? I can help test this with a massive configuration if you'd like!