@virgil-serbanuta 2024-04-25
FWIW, from the PR description it seems that, if a term is simplified multiple times, it gets the same hash/identifier. If it's reasonable, it would be nice to have a way to differentiate the various simplification attempts, perhaps a counter, e.g. instead of `[term 58b79a1][simplify]` one could have `[term 58b79a1][simplify 1]` or `[term 58b79a1:5][simplify]`.