@goodlyrottenapple 2024-04-25
> FWIW, from the PR description it seems that, if a term is simplified multiple times, it gets the same hash/identifier. If it's reasonable, it would be nice to have a way to differentiate the various simplification attempts, perhaps a counter, e.g. instead of `[term 58b79a1][simplify]` one could have `[term 58b79a1][simplify 1]` or `[term 58b79a1:5][simplify]`.

From the log above, something like `[simplify][term 58b79a1][term 24777d8][function 00be500]` means we are in the simplifier, attempting to simplify term `58b79a1`, specifically, a sub-term `24777d8` to which we are trying to apply the function equation `00be500`. Further in the logs, we attempt to apply a different equation, namely `simplification 5a8ecf6`. Given the way the contexts work, we can't change the outer-most hash until we re-assemble the term and exit the simplification context